+++ lib/core/error/result.dart
@@ -1,0 +1,15 @@
+import 'failures.dart';
+
+sealed class Result<T> {
+  const Result();
+}
+
+class Success<T> extends Result<T> {
+  final T value;
+  const Success(this.value);
+}
+
+class Failure<T> extends Result<T> {
+  final AppFailure failure;
+  const Failure(this.failure);
+}